Update resource paths for Unix tool install
Use GNUInstallDirs and an extra resource dir path to adjust output
location of resource script on non-Apple unices. Use a standard
POSIX layout of bin/ and share/, adapting MACOSX_PACKAGE_LOCATION to
point to the correct location when running on non-Apple.

Bit of a hack, so look at CMake's packaging for an example of
installing a POSIX tool in a .app bundle like structure.