Build Resourceful as static/shared client program
Refactor binreloc functionality into configurable file and INTERFACE
library with source usage requirements. This allows reconfig of
header and symbol names if needed.

Add new builds of Resourceful as a static/shared library plus command
line programs that link statically/dynamically to these. This is to
test behaviour of library with resources when used dynamically and
statically. It's expected that the static linking will always resolve
to the executabel linked in to, requiring different strategies for
resource location.

TODO: Update RFLResource to use binreloc's library location
functionality.